// packages/astro-loader-github-prs/src/graphql/gen/operations.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  /** A (potentially binary) string encoded using base64. */
  Base64String: { input: any; output: any }
  /**
   * Represents non-fractional signed whole numeric values. Since the value may
   * exceed the size of a 32-bit integer, it's encoded as a string.
   */
  BigInt: { input: any; output: any }
  /** An ISO-8601 encoded date string. */
  Date: { input: any; output: any }
  /** An ISO-8601 encoded UTC date string. */
  DateTime: { input: string; output: string }
  /** A Git object ID. */
  GitObjectID: { input: any; output: any }
  /** A fully qualified reference name (e.g. `refs/heads/master`). */
  GitRefname: { input: any; output: any }
  /** Git SSH string */
  GitSSHRemote: { input: any; output: any }
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: { input: any; output: any }
  /** A string containing HTML code. */
  HTML: { input: string; output: string }
  /** An ISO-8601 encoded UTC date string with millisecond precision. */
  PreciseDateTime: { input: any; output: any }
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: { input: string; output: string }
  /** A valid x509 certificate string */
  X509Certificate: { input: any; output: any }
}
// ...
export type PullRequestState =
  /** A pull request that has been closed without being merged. */
  | 'CLOSED'
  /** A pull request that has been closed by being merged. */
  | 'MERGED'
  /** A pull request that is still open. */
  | 'OPEN'
// ...
export type GetPrsQueryVariables = Exact<{
  search: Scalars['String']['input']
  first: Scalars['Int']['input']
  cursor?: InputMaybe<Scalars['String']['input']>
}>

export type GetPrsQuery = {
  search: {
    nodes?: Array<
      | {
          id: string
          url: string
          title: string
          titleHTML: string
          number: number
          state: PullRequestState
          isDraft: boolean
          body: string
          bodyHTML: string
          bodyText: string
          createdAt: string
          mergedAt?: string | null
          author?:
            | { login: string; url: string; avatarUrl: string }
            | { login: string; url: string; avatarUrl: string }
            | { login: string; url: string; avatarUrl: string }
            | { login: string; url: string; avatarUrl: string }
            | { login: string; url: string; avatarUrl: string }
            | null
          repository: {
            name: string
            nameWithOwner: string
            url: string
            stargazerCount: number
            isInOrganization: boolean
            owner:
              | { login: string; url: string; avatarUrl: string }
              | { login: string; url: string; avatarUrl: string }
          }
        }
      | {}
      | null
    > | null
    pageInfo: { hasNextPage: boolean; endCursor?: string | null }
  }
}

export type GetPrByNumberQueryVariables = Exact<{
  owner: Scalars['String']['input']
  repo: Scalars['String']['input']
  number: Scalars['Int']['input']
}>

export type GetPrByNumberQuery = {
  repository?: {
    pullRequest?: {
      id: string
      url: string
      title: string
      titleHTML: string
      number: number
      state: PullRequestState
      isDraft: boolean
      body: string
      bodyHTML: string
      bodyText: string
      createdAt: string
      mergedAt?: string | null
      author?:
        | { login: string; url: string; avatarUrl: string }
        | { login: string; url: string; avatarUrl: string }
        | { login: string; url: string; avatarUrl: string }
        | { login: string; url: string; avatarUrl: string }
        | { login: string; url: string; avatarUrl: string }
        | null
      repository: {
        name: string
        nameWithOwner: string
        url: string
        stargazerCount: number
        isInOrganization: boolean
        owner:
          | { login: string; url: string; avatarUrl: string }
          | { login: string; url: string; avatarUrl: string }
      }
    } | null
  } | null
}

export type GetPrByIdQueryVariables = Exact<{
  id: Scalars['ID']['input']
}>

export type GetPrByIdQuery = {
  node?:
    | {
        id: string
        url: string
        title: string
        titleHTML: string
        number: number
        state: PullRequestState
        isDraft: boolean
        body: string
        bodyHTML: string
        bodyText: string
        createdAt: string
        mergedAt?: string | null
        author?:
          | { login: string; url: string; avatarUrl: string }
          | { login: string; url: string; avatarUrl: string }
          | { login: string; url: string; avatarUrl: string }
          | { login: string; url: string; avatarUrl: string }
          | { login: string; url: string; avatarUrl: string }
          | null
        repository: {
          name: string
          nameWithOwner: string
          url: string
          stargazerCount: number
          isInOrganization: boolean
          owner:
            | { login: string; url: string; avatarUrl: string }
            | { login: string; url: string; avatarUrl: string }
        }
      }
    | {}
    | null
}
// ...
export const GetPrByIdDocument = {
  kind: 'Document',
  definitions: [
    {
      kind: 'OperationDefinition',
      operation: 'query',
      name: { kind: 'Name', value: 'GetPrById' },
      variableDefinitions: [
        {
          kind: 'VariableDefinition',
          variable: { kind: 'Variable', name: { kind: 'Name', value: 'id' } },
          type: {
            kind: 'NonNullType',
            type: { kind: 'NamedType', name: { kind: 'Name', value: 'ID' } },
          },
        },
      ],
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'node' },
            arguments: [
              {
                kind: 'Argument',
                name: { kind: 'Name', value: 'id' },
                value: {
                  kind: 'Variable',
                  name: { kind: 'Name', value: 'id' },
                },
              },
            ],
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                {
                  kind: 'FragmentSpread',
                  name: { kind: 'Name', value: 'PullRequestCore' },
                },
              ],
            },
          },
        ],
      },
    },
    {
      kind: 'FragmentDefinition',
      name: { kind: 'Name', value: 'PullRequestCore' },
      typeCondition: {
        kind: 'NamedType',
        name: { kind: 'Name', value: 'PullRequest' },
      },
      selectionSet: {
        kind: 'SelectionSet',
        selections: [
          { kind: 'Field', name: { kind: 'Name', value: 'id' } },
          { kind: 'Field', name: { kind: 'Name', value: 'url' } },
          { kind: 'Field', name: { kind: 'Name', value: 'title' } },
          { kind: 'Field', name: { kind: 'Name', value: 'titleHTML' } },
          { kind: 'Field', name: { kind: 'Name', value: 'number' } },
          { kind: 'Field', name: { kind: 'Name', value: 'state' } },
          { kind: 'Field', name: { kind: 'Name', value: 'isDraft' } },
          { kind: 'Field', name: { kind: 'Name', value: 'body' } },
          { kind: 'Field', name: { kind: 'Name', value: 'bodyHTML' } },
          { kind: 'Field', name: { kind: 'Name', value: 'bodyText' } },
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'author' },
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                { kind: 'Field', name: { kind: 'Name', value: 'login' } },
                { kind: 'Field', name: { kind: 'Name', value: 'url' } },
                { kind: 'Field', name: { kind: 'Name', value: 'avatarUrl' } },
              ],
            },
          },
          {
            kind: 'Field',
            name: { kind: 'Name', value: 'repository' },
            selectionSet: {
              kind: 'SelectionSet',
              selections: [
                { kind: 'Field', name: { kind: 'Name', value: 'name' } },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'nameWithOwner' },
                },
                { kind: 'Field', name: { kind: 'Name', value: 'url' } },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'stargazerCount' },
                },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'isInOrganization' },
                },
                {
                  kind: 'Field',
                  name: { kind: 'Name', value: 'owner' },
                  selectionSet: {
                    kind: 'SelectionSet',
                    selections: [
                      { kind: 'Field', name: { kind: 'Name', value: 'login' } },
                      { kind: 'Field', name: { kind: 'Name', value: 'url' } },
                      {
                        kind: 'Field',
                        name: { kind: 'Name', value: 'avatarUrl' },
                      },
                    ],
                  },
                },
              ],
            },
          },
          { kind: 'Field', name: { kind: 'Name', value: 'createdAt' } },
          { kind: 'Field', name: { kind: 'Name', value: 'mergedAt' } },
        ],
      },
    },
  ],
} as unknown as DocumentNode<GetPrByIdQuery, GetPrByIdQueryVariables>